Adding a new response action keyword.
The goal is to trigger an action to send back and deal with the protocol the attacker is using to understand beyond alerts.
I am always frustrated by the lack of information we get when someone/a robot tries to attack a network.

This is step #1 to turn Suricata into a honeypot ;-)